chore(ui): convert story to ts
